Rename random.variableName to random.basicString and restrict output format
