+++ backend/src/random/basic_string.js
@@ -1,10 +1,9 @@
-// Generates a pseudo-random identifier matching /^[a-z_][a-z0-9_]*$/.
+// Generates a pseudo-random string matching /^[a-z]*$/.
 // Note: this uses a seeded PRNG and is not suitable for cryptographic purposes.
 
 const { defaultGenerator } = require('./default');
 
-const FIRST_CHAR_OPTIONS = 'abcdefghijklmnopqrstuvwxyz_';
-const NEXT_CHAR_OPTIONS = '0123456789abcdefghijklmnopqrstuvwxyz_';
+const CHAR_OPTIONS = 'abcdefghijklmnopqrstuvwxyz';
 
 /** @typedef {import('./seed').NonDeterministicSeed} NonDeterministicSeed */
 
@@ -14,14 +13,14 @@ const NEXT_CHAR_OPTIONS = '0123456789abcdefghijklmnopqrstuvwxyz_';
  */
 
 /**
- * Generates a random identifier matching /^[a-z_][a-z0-9_]*$/.
+ * Generates a random lowercase latin string matching /^[a-z]*$/.
  *
  * @param {Capabilities} capabilities - An object containing a random number generator.
  * @param {number} [length=16] - The length of the generated identifier. Must be a positive integer.
- * @returns {string} A random identifier of specified length.
+ * @returns {string} A random lowercase latin string of specified length.
  * @throws {TypeError} If the length is not a positive integer.
  */
-function variableName(capabilities, length = 16) {
+function basicString(capabilities, length = 16) {
     if (!Number.isInteger(length) || length < 1) {
         throw new TypeError('Length must be a positive integer');
     }
@@ -29,16 +28,13 @@ function variableName(capabilities, length = 16) {
     const rng = defaultGenerator(capabilities);
     const result = new Array(length);
 
-    const firstIdx = rng.nextInt(0, FIRST_CHAR_OPTIONS.length - 1);
-    result[0] = FIRST_CHAR_OPTIONS[firstIdx];
-
-    const nextCharLen = NEXT_CHAR_OPTIONS.length;
-    for (let i = 1; i < length; i++) {
-        const idx = rng.nextInt(0, nextCharLen - 1);
-        result[i] = NEXT_CHAR_OPTIONS[idx];
+    const charLen = CHAR_OPTIONS.length;
+    for (let i = 0; i < length; i++) {
+        const idx = rng.nextInt(0, charLen - 1);
+        result[i] = CHAR_OPTIONS[idx];
     }
 
     return result.join('');
 }
 
-module.exports = { variableName };
+module.exports = { basicString };
